+++ src/connections/InitWeights.cpp
@@ -225,7 +225,7 @@ int InitWeights::gauss2DCalcWeights(PVPatch * patch, InitGauss2DWeightsParams * 
    int sx_tmp=weightParamPtr->getsx_tmp();
    int sy_tmp=weightParamPtr->getsy_tmp();
    int sf_tmp=weightParamPtr->getsf_tmp();
-   float r2Max=weightParamPtr->getr2Max();
+   double r2Max=weightParamPtr->getr2Max();
 
    pvdata_t * w_tmp = patch->data;
